third_party/WebKit/LayoutTests/fast/dom/interface-object-proto-expected.txt

Tests the prototype chain of interface objects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Object.getPrototypeOf(HTMLBodyElement) is HTMLElement
PASS Object.getPrototypeOf(HTMLElement) is Element
PASS Object.getPrototypeOf(Element) is Node
PASS Object.getPrototypeOf(Node) is EventTarget
PASS Object.getPrototypeOf(EventTarget) is Function.prototype
PASS successfullyParsed is true

TEST COMPLETE
